remove deprecated code
