--- src/components/AppBar/Burger.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import { StyledBurger } from './AppBar.style';
-
-const Burger = ({
-  open,
-  setOpen,
-}: {
-  open: boolean;
-  setOpen: (open: boolean) => void;
-}) => {
-  return (
-    <StyledBurger open={open} onClick={() => setOpen(!open)}>
-      <div />
-      <div />
-      <div />
-    </StyledBurger>
-  );
-};
-
-export default Burger;
